+++ packages/did-ui-react/src/utils/telegram.ts
@@ -5,7 +5,6 @@ import { pushEncodeMessage, pushMessageByApi } from './openlogin/crossTabMessage
 import { decodeMessageByRsaKey, did, handleErrorMessage } from '.';
 import qs from 'query-string';
 import { TelegramWebappInitData } from '@portkey/types';
-import { PORTKEY_LOGIN_STORAGE_KEY } from '../constants/storage';
 import { getServiceUrl } from '../components/config-provider/utils';
 
 export function getTelegram() {
@@ -66,71 +65,9 @@ export function openLinkFromTelegram(url: string, params: Record<string, any>) {
   }
 }
 
-export async function saveDataWithInTelegram({
-  needPersist,
-  isSaveDataToStorage = true,
-  isOpenTelegramLink = true,
-  loginId = '',
-  storageKey = '',
-  storageValue,
-  pushMessage,
-  telegramLink,
-  onBeforeOpenLink,
-}: {
-  needPersist: boolean;
-  isSaveDataToStorage?: boolean;
-  isOpenTelegramLink?: boolean;
-  loginId?: string;
-  storageKey?: string;
-  storageValue?: Record<string, any>;
-  pushMessage?: Record<string, any>;
-  telegramLink?: string;
-  onBeforeOpenLink?: () => Promise<void> | void;
-}) {
-  // Save Data
-  // 1. Get publicKey and privateKey
-  const cryptoManager = new forgeWeb.ForgeCryptoManager();
-  const keyPair = await cryptoManager.generateKeyPair();
-  loginId = loginId || randomId();
-  const sessionAuth = JSON.stringify({
-    loginId: loginId,
-    publicKey: keyPair.publicKey,
-    needPersist,
-  });
-
-  // 2. Save data to storage, default save privateKey
-  if (isSaveDataToStorage) {
-    let storageValueConcat = { rsaKey: keyPair.privateKey, loginId };
-    if (storageValue && typeof storageValue === 'object') {
-      storageValueConcat = Object.assign({}, storageValueConcat, storageValue);
-    }
-    await did.config.storageMethod.setItem(storageKey, JSON.stringify(storageValueConcat));
-  }
-
-  // 3. Save data to database, default save storageKey
-  let params = { storageKey };
-  if (pushMessage && typeof pushMessage === 'object') {
-    params = Object.assign({}, params, pushMessage);
-  }
-  await pushEncodeMessage(sessionAuth, CrossTabPushMessageType.onAuthStatusChanged, JSON.stringify(params));
-
-  await onBeforeOpenLink?.();
-
-  // 4. Open telegram link
-  if (isOpenTelegramLink && telegramLink) {
-    const Telegram = getTelegram();
-    Telegram?.WebApp.openTelegramLink(`${telegramLink}?startapp=${loginId}`);
-  }
-  return {
-    loginId: loginId,
-    publicKey: keyPair.publicKey,
-    rsaKey: keyPair.privateKey,
-  };
-}
-
 // usage: telegram auth in dapp-webapp
 export async function saveEncodeInfoToStorageAndPortkeyDatabase(
-  storageKey: string,
+  // storageKey: string,
   methodName: CrossTabPushMessageType,
   extraData?: Record<string, any>,
 ) {
@@ -141,11 +78,10 @@ export async function saveEncodeInfoToStorageAndPortkeyDatabase(
 
   // 2. Save encode info to dapp localStorage
   const storageValue = {
-    loginId,
     rsaKey: keyPair.privateKey,
     methodName,
   };
-  await did.config.storageMethod.setItem(storageKey, JSON.stringify(storageValue));
+  await did.config.storageMethod.setItem(loginId, JSON.stringify(storageValue));
   let concatData = { publicKey: keyPair.publicKey };
   if (extraData && typeof extraData === 'object') {
     concatData = Object.assign({}, concatData, extraData);
@@ -201,11 +137,10 @@ export async function saveAccessTokenToPortkeyDatabase(
 }
 
 // usage: back to dapp-webapp
-export async function getPrivateKeyFromLocalStorage(storageKey: string, matchLoginId: string) {
-  const storage = await did.config.storageMethod.getItem(storageKey);
+export async function getPrivateKeyFromLocalStorage(loginId: string) {
+  const storage = await did.config.storageMethod.getItem(loginId);
   if (storage && typeof storage === 'string' && storage.length > 0) {
-    const { rsaKey, loginId } = JSON.parse(storage);
-    return loginId === matchLoginId ? rsaKey : null;
+    return JSON.parse(storage)?.rsaKey;
   }
   return null;
 }
@@ -220,12 +155,9 @@ export async function getAndDecodeAccessToken(loginId: string, methodName: Cross
 // usage: first step - attach telegram login in dapp-webapp
 export async function saveDataAndOpenPortkeyWebapp(yourTelegramLink: string, targetTelegramLink: string) {
   try {
-    const storageKey = `${PORTKEY_LOGIN_STORAGE_KEY}_Telegram`;
-    const { loginId } = await saveEncodeInfoToStorageAndPortkeyDatabase(
-      storageKey,
-      CrossTabPushMessageType.onSavePublicKey,
-      { yourTelegramLink },
-    );
+    const { loginId } = await saveEncodeInfoToStorageAndPortkeyDatabase(CrossTabPushMessageType.onSavePublicKey, {
+      yourTelegramLink,
+    });
     console.log('=== loginId', loginId);
     if (targetTelegramLink) {
       const Telegram = getTelegram();
@@ -284,8 +216,7 @@ export async function getAccessTokenAndOpenPortkeyWebapp({
 // usage: last step - get accessToken in dapp-webapp
 export async function getAccessTokenInDappTelegram(loginId: string) {
   try {
-    const storageKey = `${PORTKEY_LOGIN_STORAGE_KEY}_Telegram`;
-    const rsaKey = await getPrivateKeyFromLocalStorage(storageKey, loginId);
+    const rsaKey = await getPrivateKeyFromLocalStorage(loginId);
     console.log('=== rsaKey', rsaKey);
     if (!rsaKey) return; // TODO tg
     return await getAndDecodeAccessToken(loginId, CrossTabPushMessageType.onAuthStatusChanged, rsaKey);