+++ packages/next-example/src/app/assets/page.tsx
@@ -1,6 +1,7 @@
 'use client';
 
 import { ConfigProvider, Asset, PortkeyAssetProvider } from '@portkey/did-ui-react';
+import { Button } from 'antd';
 import { useRouter } from 'next/navigation';
 import React from 'react';
 
@@ -21,6 +22,9 @@ export default function Assets() {
   const router = useRouter();
   return (
     <PortkeyAssetProvider pin="111111" originChainId="AELF">
+      <a href="dapp-webapp">
+        <Button>Go to dapp-webapp</Button>
+      </a>
       <Asset
         faucet={{
           faucetContractAddress: '2UM9eusxdRyCztbmMZadGXzwgwKfFdk8pF4ckw58D769ehaPSR',